Added code documentation
I added code documentation for all the utility classes in the portfolio. I also added documentation to the home controller, and moved the Alert.cs file to the utilities folder.

+++ Portfolio/Portfolio/Utilities/CalculateTotal.cs
@@ -2,8 +2,16 @@
 
 namespace Portfolio.Utilities
 {
+    /// <summary>
+    /// Used for calculating prices in the 4th Wall Café application.
+    /// </summary>
     public static class CalculateTotal
     {
+        /// <summary>
+        /// Takes a list of shopping bag items and calculates the total.
+        /// </summary>
+        /// <param name="items">A list of shopping bag items.</param>
+        /// <returns>The total price of all items in decimal form.</returns>
         public static decimal AddItems(List<ShoppingBagItemResponse> items)
         {
             decimal total = 0;
@@ -15,7 +23,7 @@ namespace Portfolio.Utilities
 
                 var itemTotal = price * quantity;
 
-                total += (decimal)itemTotal;
+                total += itemTotal;
             }
 
             return total;
